fix timer and add task, but multi task crashed, so sad

+++ inc/lib/string.h
@@ -1,16 +1,15 @@
 #ifndef __STRING_H__
 #define __STRING_H__
 
-#include "sys/types.h"
 
 char *itoa (int val, int base);
-int strcmp(const char* str1, const char* str2);
+int  strcmp(const char* str1, const char* str2);
 void strcpy(char *s, char *t);
-size_t strlen(const char *str);
+int  strlen(const char *str);
 
-void *memcpy(void *dest, const void *src, size_t count);
-void *memset(void *dest, char val, size_t count);
-unsigned short *memsetw(unsigned short *dest, unsigned short val, size_t count);
+void *memcpy(void *dest, const void *src, int count);
+void *memset(void *dest, char val, int count);
+unsigned short *memsetw(unsigned short *dest, unsigned short val, int count);
 
 char *strtok(char * s,const char * ct);
 char *strpbrk(const char * cs,const char * ct);
