+++ src/lib/string.c
@@ -35,14 +35,14 @@ void strcpy(char *s, char *t)
 	}
 }
 
-size_t strlen(const char *str)
+int strlen(const char *str)
 {
-    size_t retval;
+    int retval;
     for(retval = 0; *str != '\0'; str++) retval++;
     return retval;
 }
 
-void *memcpy(void *dest, const void *src, size_t count)
+void *memcpy(void *dest, const void *src, int count)
 {
     const char *sp = (const char *)src;
     char *dp = (char *)dest;
@@ -50,14 +50,14 @@ void *memcpy(void *dest, const void *src, size_t count)
     return dest;
 }
 
-void *memset(void *dest, char val, size_t count)
+void *memset(void *dest, char val, int count)
 {
     char *temp = (char *)dest;
     for( ; count != 0; count--) *temp++ = val;
     return dest;
 }
 
-unsigned short *memsetw(unsigned short *dest, unsigned short val, size_t count)
+unsigned short *memsetw(unsigned short *dest, unsigned short val, int count)
 {
     unsigned short *temp = (unsigned short *)dest;
     for( ; count != 0; count--) *temp++ = val;
